Scheduling settings moved to a new settings file
