+++ app/src/main/res/values/strings.xml
@@ -99,4 +99,6 @@
     <string name="pref_stop_info_stop_start">Stop:</string>
     <string name="pref_stop_info_stop_battery">low battery</string>
     <string name="pref_stop_info_stop_storage">low storage</string>
+    <string name="pref_scheduling_settings_category">Scheduling</string>
+    <string name="pref_scheduling_settings_title">Configure Scheduling</string>
 </resources>
